Move checkEndScreen to TypinGameActivityTest
